rename enumerable folder to stream
