+++ packages/punchcard/lib/index.ts
@@ -1,6 +1,6 @@
 export * from './api-gateway';
 export * from './data-lake';
-export * from './enumerable';
+export * from './stream';
 export * from './compute';
 export * from './storage';
 export * from './shape';
