+++ packages/punchcard/lib/stream/index.ts
@@ -1,8 +1,8 @@
 export * from './collector';
 export * from './delivery-stream';
-export * from './enumerable';
 export * from './glue';
+export * from './kinesis';
 export * from './queue';
 export * from './sink';
-export * from './kinesis';
+export * from './stream';
 export * from './topic';
