+++ packages/punchcard/lib/stream/resource.ts
