+++ packages/punchcard/lib/stream/sink.ts
